refactor to db_store package
