+++ server/.gitignore
@@ -22,4 +22,8 @@ go.work
 go.work.sum
 
 # env file
+.env
 *.env
+
+# others
+data
