+++ server/infra/store/db_store/internal/gen_jet/skyvault/public/model/shares.go
