+++ server/infra/store/store.go
@@ -1,99 +1,22 @@
 package store
 
 import (
-	"context"
-	"fmt"
-	"log"
-	"log/slog"
-	"os"
-	"skyvault/common"
-	"time"
-
-	"github.com/golang-migrate/migrate/v4"
-	"github.com/golang-migrate/migrate/v4/database/postgres"
-	_ "github.com/golang-migrate/migrate/v4/source/file"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/jackc/pgx/v5/stdlib"
+	"skyvault/domain/auth"
+	"skyvault/infra/store/db_store"
 )
 
+var _ auth.Repo = &BaseAuthRepo{}
+
 type Store struct {
-	db *dbStore
+	db *db_store.DBStore
 }
 
-func NewStore(dbStore *dbStore) *Store {
+func NewStore(dbStore *db_store.DBStore) *Store {
 	return &Store{
 		db: dbStore,
 	}
 }
 
-type dbStore struct {
-	*pgxpool.Pool
-}
-
-func NewDBStore(connStr string) *dbStore {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer func() {
-		cancel()
-	}()
-
-	pool, err := connectDatabase(connStr)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	err = pool.Ping(ctx)
-	if err != nil {
-		slog.ErrorContext(ctx, "failed to ping the db", slog.Any("error", err))
-		os.Exit(1)
-	}
-
-	return &dbStore{pool}
-}
-
-func connectDatabase(dsn string) (*pgxpool.Pool, error) {
-	config, err := pgxpool.ParseConfig(dsn)
-	if err != nil {
-		return nil, fmt.Errorf("failed to parse dsn config: %w", err)
-	}
-
-	pool, err := pgxpool.NewWithConfig(context.Background(), config)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create pool: %w", err)
-	}
-
-	return pool, nil
-}
-
-func (s *dbStore) MigrateUp() {
-	stdDB := stdlib.OpenDBFromPool(s.Pool)
-	defer func() {
-		if err := stdDB.Close(); err != nil {
-			log.Fatal("failed to close the std DB opened from the pool", err)
-		}
-	}()
-
-	driver, err := postgres.WithInstance(stdDB, &postgres.Config{})
-	if err != nil {
-		log.Fatal("failed to create postgres driver", err)
-	}
-	defer func() {
-		if err := driver.Close(); err != nil {
-			log.Fatal("failed to close the postgres driver", err)
-		}
-	}()
-
-	testMigrate, err := migrate.NewWithDatabaseInstance(fmt.Sprintf("file://%s", common.Configs.DB_MIGRATION_PATH), "postgres", driver)
-	if err != nil {
-		log.Fatal("failed to create migrate instance", err)
-	}
-	err = testMigrate.Up()
-	if err != nil {
-		log.Fatal("failed to migrate up", err)
-	}
-}
-
-func (s *Store) NewAuthRepo() *AuthRepo {
-	return &AuthRepo{
-		db: s.db,
-	}
+type BaseAuthRepo struct {
+	db_store.AuthRepo
 }
